soma: move soma.Config into its own package
The supervisor needs access to the SOMA runtime configuration, so
it requires a pointer to soma.SOMA. This requires an import of the
package.
soma.SOMA needs to import package super to start super.Supervisor,
this would lead to cyclic import (which go does not like).

Move the configuration file into its own package, this way both
soma and super can import it without issue.