super: always use server-supplied IV during kex
The server replacing the client-provided IV was always
part of the protocol, but has now been promoted to
default case.
A client providing an IV not considered a bad request,
the value is simply ignored.

Abort if generation or validation of the IV does not
generate a usable IV in a small finite number of attempts.
This likely means the entropy source has gone mad,
making it unlikely to recover in another n++ attempts.