super: cleanup Unauthorized vs Forbidden
Only use Unauthorized as the result of failed basic auth,
since this also sets the RFC required realm headers.
Otherwise, have the supervisor always return Forbidden. No
error messages are transported to the client, they are only
logged server side.

The exception is Conflict, which is used to transport a policy
violation to the user.