auth: add functions to update the IV of kex
While there is a function to re-generate the IV, there is no helper
for the peer to export/import the re-generated IV into the local
view of the KEX.

This caused the underlying cryptography to work and prevent decryption
of data encrypted with IV A and decrypted with IV B. Very good, but
reduces practical application.

Add the missing helper so the client can import the new IV if the
server chose to re-generate it.