msg: IsOK() is false for set errors
If the request was successful but did not actually do anything
the created result is 200/OK with Error==No rows affected.

In that case, whatever request was received did not have an effect
and should not be used to update the supervisor cache. IsOK() is
used to decide if the supervisor cache requires an update, so
return false whenever Error != nil.