proto: update Job, add JobType|JobResult|JobStatus
Database schema soma.201811120002 upgraded the job related
database tables, which contained the special tables
- soma.job_type
- soma.job_status
- soma.job_result
which were the only three tables whose content was solely
managed via somadbctl and not exposed as part of the API
protocol.
This meant that additional database rows required a schema
upgrade.

Update the protocol library with structs for these tables
so that they can be migrated to soma supported endpoints.