rest: add and use Rest.hardConflict()
hardConflict returns a 409 HTTP error with a provided error text.
This function is intended to be used only for special policy
violations.

It is currently used by the SOMA client during account activation.
Once the client code has been adapted, there is no reason why this
error can not be returned in the 'normal' application data error
format.