FIX: schema reference for pgcrypto extension
This commit cherry-picks from the following two commits, but uses public
in all locations instead of installing into a postgresql system table.

- 2adf2003fc1d4a803cc1619243cc12f2a883da73
fixed schema for  extension pgcrypto

Seems, if extension is put in  schema public, all calls to gen_random_uuid need to be
prefixed with 'public.'. Put it in pg_catalog, and prefix is not
needed

- 37e5c7fd828eae5d566bdbda438443b9673a2205
more comma/namespace errors