tree: risk runtime panic to shut up stupid linter
No, the sync.RWMutex is not copied. The clone gets a new lock
because the zero value of sync.RWMutex is a new, ready-to-use
lock.

Now, it is *sync.RWMutex. The zero value is now a nilptr, using
it is a fatal runtime crash. Now the world is not better. But the
stupid linters of this world no longer complain, because now
the data type _LOOKS_ like the lock is passed by reference when
it is not and has never been passed around AT FUCKING ALL.